requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-json-logger==2.0.7
prometheus-client==0.20.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests